Add ability to specify amount and balance filters
