+++ Web/Models/Viia/ViiaQueryBetweenValuesModel.cs
@@ -1,0 +1,8 @@
+namespace ViiaSample.Models.Viia
+{
+    public class ViiaQueryBetweenValuesModel
+    {
+        public double Min { get; set; }
+        public double Max { get; set; }
+    }
+}
